Refactor: form 구간 components로 세분화
