+++ moodoodleFront/moodoodlefront/src/App.js
@@ -18,7 +18,7 @@ function App() {
             <Pc className='flex flex-col m-auto'>
               <div className='flex-1'>
                 <Routes>
-                  <Route exact path='/' element={<Start />}>
+                  <Route exact path='/' element={<Home />}>
                     <Route path='/' element={<Main />} />
                     <Route
                       path='/diary/:selectedDate'
